Cleanups so it doesn't need a dist dir
