--- packages/fixie/index.ts
@@ -1,2 +1,0 @@
-export { FixieClient, User, FixieClientError } from './src/client.js';
-export * from './src/sidekick-types.js';
