+++ packages/fixie/package.json
@@ -1,6 +1,6 @@
 {
   "name": "fixie",
-  "version": "4.1.0",
+  "version": "4.1.1",
   "license": "MIT",
   "repository": "fixie-ai/ai-jsx",
   "bugs": "https://github.com/fixie-ai/ai-jsx/issues",
@@ -8,7 +8,7 @@
   "type": "module",
   "scripts": {
     "build": "rm -rf dist && tsc",
-    "start": "node --no-warnings src/main.js",
+    "start": "node --no-warnings dist/main.js",
     "build-start": "yarn run build && yarn run start",
     "format": "prettier --write .",
     "test": "yarn run build && yarn run lint",
@@ -19,7 +19,9 @@
   "volta": {
     "extends": "../../package.json"
   },
-  "bin": "./src/main.js",
+  "bin": "./dist/main.js",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
   "dependencies": {
     "@apollo/client": "^3.8.1",
     "@types/apollo-upload-client": "^17.0.2",
