+++ packages/fixie/src/index.ts
@@ -1,0 +1,3 @@
+export * from './client.js';
+export * from './fixie-embed.js';
+export * from './types.js';
