--- packages/fixie/web.ts
@@ -1,8 +1,0 @@
-export { FixieClient } from './src/client.js';
-export {
-  InlineFixieEmbed,
-  FloatingFixieEmbed,
-  ControlledFloatingFixieEmbed,
-  getBaseIframeProps,
-} from './src/fixie-embed.js';
-export * from './src/use-fixie.js';
